server/django/requirements/base.txt
Django==4.2.16
psycopg[binary]
Pillow
django-rest-framework
djangorestframework-simplejwt
djoser
django-cors-headers
django-mptt
num2words
django-separatedvaluesfield
django-xhtml2pdf
djangorestframework-jwt
django-filter
django-ses
requests
django-import-export
django-auditlog
django-q
redis
openpyxl==3.0.10
XlsxWriter
whitenoise
thefuzz
django-dbbackup
django-storages[boto3]
